android: Do not use queryTxtRecord
Current implementation of queryTxtRecord issues serviceLost and
serviceFound events when TXT records are resolved right after the
service has been found.

This commit disables this functionality so that serviceLost and
serviceFound events represent real state of MDNS network. We can do this
because we do not rely on the resolved TXT records - we resolve the
service right after it was found which also resolve TXT records.

Signed-off-by: Daniel Jenikovský <daniel.jenikovsky@streamunlimited.com>